add monitor emq

+++ assets/vehicle.sql
@@ -521,9 +521,10 @@ CREATE TABLE IF NOT EXISTS `monitors`  (
   `deleted_at` timestamp NULL DEFAULT NULL,
 
   `monitor_id` varchar(255)  NULL DEFAULT NULL,
-  `cpu_rate` int(3)  NULL DEFAULT NULL,
-  `mem_rate` int(3)  NULL DEFAULT NULL,
+  `cpu_rate` double  NULL DEFAULT NULL,
+  `mem_rate` double  NULL DEFAULT NULL,
   `gather_time` int(11) NULL DEFAULT NULL,
+
 
   PRIMARY KEY (`id`) USING BTREE,
   INDEX `idx_monitors_deleted_at`(`deleted_at`) USING BTREE
@@ -538,7 +539,7 @@ CREATE TABLE IF NOT EXISTS `disks`  (
 
   `monitor_id` varchar(255)  NULL DEFAULT NULL,
   `path` varchar(255)  NULL DEFAULT NULL,
-  `disk_rate` int(3)  NULL DEFAULT NULL,
+  `disk_rate` double  NULL DEFAULT NULL,
 
   PRIMARY KEY (`id`) USING BTREE,
   INDEX `idx_disks_deleted_at`(`deleted_at`) USING BTREE
